+++ idf_component_manager/utils.py
@@ -1,11 +1,18 @@
 # SPDX-FileCopyrightText: 2022-2023 Espressif Systems (Shanghai) CO LTD
 # SPDX-License-Identifier: Apache-2.0
+
+import enum
 import re
 
 import click
 from packaging import version
 
 from idf_component_tools.messages import UserHint, UserNotice
+
+try:
+    from functools32 import lru_cache
+except ImportError:
+    from functools import lru_cache
 
 CLICK_SUPPORTS_SHOW_DEFAULT = version.parse(click.__version__) >= version.parse('7.1.0')
 
@@ -59,3 +66,65 @@ def print_info(
 
 
 RE_PATTERN = type(re.compile(''))  # this is a workaround for `re.Pattern` for python<3.7
+
+
+# total_ordering will raise an error in python 2.7 with enum34
+#   ValueError: must define at least one ordering operation: < > <= >=
+# The reason is that the `dir()` behavior is different,
+# ___eq__, __lt__, __hash__ are not in the dir() result.
+# define all six operators here
+class ComponentType(str, enum.Enum):
+    # These double-quotes are coming from the build system
+    IDF_COMPONENTS = '"idf_components"'
+    PROJECT_MANAGED_COMPONENTS = '"project_managed_components"'
+    PROJECT_COMPONENTS = '"project_components"'
+    PROJECT_EXTRA_COMPONENTS = '"project_extra_components"'
+
+    # the lower value is, the lower priority it is
+    @classmethod
+    def order(cls):
+        return {
+            cls.IDF_COMPONENTS: 0,
+            cls.PROJECT_MANAGED_COMPONENTS: 1,
+            cls.PROJECT_COMPONENTS: 2,
+            cls.PROJECT_EXTRA_COMPONENTS: 3,
+        }
+
+    def __hash__(self):
+        return hash(self.value)
+
+    def __eq__(self, other):
+        if not isinstance(other, ComponentType):
+            return NotImplemented
+
+        return self.value == other.value
+
+    def __ne__(self, other):
+        if not isinstance(other, ComponentType):
+            return NotImplemented
+
+        return self.value != other.value
+
+    def __lt__(self, other):
+        if not isinstance(other, ComponentType):
+            return NotImplemented
+
+        return self.order()[self] < self.order()[other]
+
+    def __le__(self, other):
+        if not isinstance(other, ComponentType):
+            return NotImplemented
+
+        return self < other or self == other
+
+    def __gt__(self, other):
+        if not isinstance(other, ComponentType):
+            return NotImplemented
+
+        return not self <= other
+
+    def __ge__(self, other):
+        if not isinstance(other, ComponentType):
+            return NotImplemented
+
+        return not self < other